Built a script to auto-generate all modules pages and a table of links to modules pages.
This is a big commit. I created a Jekyll plugin called pullModules.rb to pull all our repo READMEs and screenshots, compile them into site.pages instances, and create a table on the /modules/ page that links to them directly.

To pull the repos, I needed to install a few gems that I added to the Gemfile. Note that github won't run user-made plugins, so to push to git, push the contents of the _site directory and include the file ".nojekyll" in its root. The layout of each module page is from _layout/module.html

To make the table, I created another layout in _layout/module_index.html. The table itself uses the DataTables jQuery plugin, and its contents were generated using Liquid tags in _include/moduletable.html.

Note that at this point, this may be the longest commit in RTXI's commit history, but I must admit it is still woefully inadequate. I'm going to create a new branch in the rtxi.github.io repo for the base of the build directory (what is currently in master), and master will from now on contain the contents of the _site directory. (Look in future for me to make a rakefile to handle this. Far away in the future.)

-Ansel