Created a page for displaying papers and added all publications from rtxi.org
Each publication is a markdown-formatted post in the _posts/papers directory. To display them, I created a new layout in _layouts/paper.html, and for papers/index.html, I used Liquid tags to create a loop that displays each post sorted by year.

I still need to find a way to not display ALL of the publications at once (because the page is absurdly cluttered right now), and make use the the masonry (?) jQuery plugin to make each panel for each publication align better so there aren't any weird gaps. Doing that with default bootsrap is not supported.

That's all. Cheers.

-Ansel