rebuild
Note: all changes to the master branch are just offshoots of whatever was changed in the build branch. Therefore, all changes in each commit on the master branch are more-or-less reflected in the commit messages for the build branch. From now on, all the commit messages here with say "rebuild" or something else that short and lacking in detail. Besides, looking through the master branch, which just contains `bundle exec jekyll serve` output, to make changes defeats the purpose of using HTML generators like Jekyll.